Breaks the ATL code into smaller files
